refactor to remove authorizing presentation from assuranceSession
