+++ AEPAssurance/Source/Assurance.swift
@@ -122,15 +122,16 @@ public class Assurance: NSObject, Extension {
     private func handleAssuranceRequestContent(event: Event) {
         /// early bail out if eventData is nil
         guard let startSessionData = event.data else {
-            #if DEBUG
-            if event.name == AssuranceConstants.AssuranceEvent.Name.QUICKCONNECT_START_SESSION {
-                sessionOrchestrator.createQuickConnectSession()
-                return
-            }
-            #endif
             Log.debug(label: AssuranceConstants.LOG_TAG, "Assurance start session event received with empty data. Dropping event.")
             return
         }
+        
+        #if DEBUG
+        if let isQuickConnect = startSessionData[AssuranceConstants.EventDataKey.QUICK_CONNECT] as? Bool, isQuickConnect {
+            sessionOrchestrator.startQuickConnectFlow()
+            return
+        }
+        #endif
 
         guard let deeplinkUrlString = startSessionData[AssuranceConstants.EventDataKey.START_SESSION_URL] as? String else {
             Log.debug(label: AssuranceConstants.LOG_TAG, "Assurance start session event received with no deeplink url. Dropping event.")
@@ -168,9 +169,9 @@ public class Assurance: NSObject, Extension {
     ///     - event - a mobileCore's places request event
     private func handlePlacesRequest(event: Event) {
         if event.isRequestNearByPOIEvent {
-            sessionOrchestrator.session?.presentation.addClientLog("Places - Requesting \(event.poiCount) nearby POIs from (\(event.latitude), \(event.longitude))", visibility: .normal)
+            sessionOrchestrator.session?.statusPresentation.addClientLog("Places - Requesting \(event.poiCount) nearby POIs from (\(event.latitude), \(event.longitude))", visibility: .normal)
         } else if event.isRequestResetEvent {
-            sessionOrchestrator.session?.presentation.addClientLog("Places - Resetting location", visibility: .normal)
+            sessionOrchestrator.session?.statusPresentation.addClientLog("Places - Resetting location", visibility: .normal)
         }
     }
 
@@ -180,16 +181,16 @@ public class Assurance: NSObject, Extension {
     ///     - event - a mobileCore's places response event
     private func handlePlacesResponse(event: Event) {
         if event.isResponseRegionEvent {
-            sessionOrchestrator.session?.presentation.addClientLog("Places - Processed \(event.regionEventType) for region \(event.regionName).", visibility: .normal)
+            sessionOrchestrator.session?.statusPresentation.addClientLog("Places - Processed \(event.regionEventType) for region \(event.regionName).", visibility: .normal)
         } else if event.isResponseNearByEvent {
             let nearByPOIs = event.nearByPOIs
             for poi in nearByPOIs {
                 guard let poiDictionary = poi as? [String: Any] else {
                     return
                 }
-                sessionOrchestrator.session?.presentation.addClientLog("\t  \(poiDictionary["regionname"] as? String ?? "Unknown")", visibility: .high)
+                sessionOrchestrator.session?.statusPresentation.addClientLog("\t  \(poiDictionary["regionname"] as? String ?? "Unknown")", visibility: .high)
             }
-            sessionOrchestrator.session?.presentation.addClientLog("Places - Found \(nearByPOIs.count) nearby POIs\(!nearByPOIs.isEmpty ? " :" : ".")", visibility: .high)
+            sessionOrchestrator.session?.statusPresentation.addClientLog("Places - Found \(nearByPOIs.count) nearby POIs\(!nearByPOIs.isEmpty ? " :" : ".")", visibility: .high)
         }
     }
